feat: getInstanceBoardId
